+++ electronics_abstract_parts/AbstractResistor.py
@@ -145,66 +145,3 @@ class CurrentSenseResistor(DiscreteApplication):
     self.connect(self.pwr_in, self.pwr_adapter.src)
     self.connect(self.pwr_adapter.dst, self.sense_in)
     self.connect(self.res.pwr_out.as_analog_source(), self.sense_out)
-
-
-@abstract_block
-class UnpolarizedCapacitor(PassiveComponent):
-  """Base type for a capacitor, that defines its parameters and without ports (since capacitors can be polarized)"""
-  @init_in_parent
-  def __init__(self, capacitance: RangeLike, voltage: RangeLike) -> None:
-    super().__init__()
-
-    self.capacitance = self.ArgParameter(capacitance)
-    self.voltage = self.ArgParameter(voltage)  # defined as operating voltage range
-
-
-@abstract_block
-class Capacitor(UnpolarizedCapacitor):
-  """Polarized capacitor, which we assume will be the default"""
-  @init_in_parent
-  def __init__(self, *args, **kwargs) -> None:
-    super().__init__(*args, **kwargs)
-
-    self.pos = self.Port(Passive.empty())
-    self.neg = self.Port(Passive.empty())
-
-
-class DecouplingCapacitor(DiscreteApplication):
-  """Optionally polarized capacitor used for DC decoupling, with VoltageSink connections with voltage inference.
-  Implemented as a shim block."""
-  @init_in_parent
-  def __init__(self, capacitance: RangeLike) -> None:
-    super().__init__()
-
-    self.cap = self.Block(Capacitor(capacitance, voltage=RangeExpr()))
-    self.gnd = self.Export(self.cap.neg.as_voltage_sink(), [Common])
-    self.pwr = self.Export(self.cap.pos.as_voltage_sink(), [Power])
-
-    self.assign(self.cap.voltage, self.pwr.link().voltage - self.gnd.link().voltage)
-
-  def connected(self, gnd: Optional[Port[VoltageLink]] = None, pwr: Optional[Port[VoltageLink]] = None) -> \
-      'DecouplingCapacitor':
-    """Convenience function to connect both ports, returning this object so it can still be given a name."""
-    if gnd is not None:
-      cast(Block, builder.get_enclosing_block()).connect(gnd, self.gnd)
-    if pwr is not None:
-      cast(Block, builder.get_enclosing_block()).connect(pwr, self.pwr)
-    return self
-
-
-@abstract_block
-class Inductor(PassiveComponent):
-  @init_in_parent
-  def __init__(self, inductance: RangeLike,
-               current: RangeLike = Default(RangeExpr.ZERO),
-               frequency: RangeLike = Default(RangeExpr.EMPTY_ZERO)) -> None:
-    super().__init__()
-
-    self.a = self.Port(Passive.empty())
-    self.b = self.Port(Passive.empty())
-
-    self.inductance = self.ArgParameter(inductance)
-    self.current = self.ArgParameter(current)  # defined as operating current range, non-directioned
-    self.frequency = self.ArgParameter(frequency)  # defined as operating frequency range
-    # TODO: in the future, when we consider efficiency - for now, use current ratings
-    # self.resistance_dc = self.Parameter(RangeExpr())
